Minor wordsmithing to reflect current practice
Generally things are done by consensus, with the core team stepping in when
there's a dispute. Update the documented encumbered software policy to reflect
what's actually done and is considered to be working. This harmonizes the wording
for contributed software and encumbered software. Also wordsmith how Release
Engineering is referenced. Also removes stray > marks on object files. Also
document the fw files.

Note: the ascii file requierment dates from CVS days and should change. But
I've not pushed things that far in this commit.