Update shared library advice now that a.out no longer is a consideration
Remove old a.out information. Update the policy to match our current practice of
rarely, if ever, bumping the major version of libraries, but instead relying on
symbol versioning to maintain a stable ABI. Move from paragraph form to bullet
points. Also, tweak the 'once per major release' policy to 'as needed' because
the conditions that existed for the former no longer apply.

Discussed with: jhb@
Reviewed by: kib@, debdrub@
Differential Revision: https://reviews.freebsd.org/D28486