Update relnotes.adoc
Add release notes for bin/calendar, usr.bin/gh-bc (contrib/bc), and
the new getlocalbase() function in libutil

This is the first commit regarding actual changes of base components
in this document and I'm not sure about the style and format expected.
Feel free to edit and adapt the text from this commit (or let me know
how it should be changed) to make it better align with later additions
to this file.

Approved by:	re