FDP: Refer to the proper package name for hugo
The package name 'hugo' refers to emulators/hugo, a PC Engine
(TurboGrafx 16) emulator, whereas the hugo that is required by the new
framework is www/gohugo. A similar case applies with the package name of
AsciiDoctor, which is 'rubygem-asciidoctor'.
Therefore fix package names so they refer to proper ones.

PR:		253470