Makefiles: Add a conditional check on HOSTNAME variable
In order for hugo to build all links correctly, it is necessary to
define baseURL - but in order to do this, we need some way of adding a
hostname. This change adds the HOSTNAME variable to accomplish this.

Another advantage is that if HOSTNAME is set to something other than
localhost, hugo can be run on a different (more powerful) machine which
can shorten the time it takes to build.

A note about this will need to be added to the FDP later.