Makefiles and FDP: Change python binary
Currently '/usr/local/bin/python3.7' is used as the value of
PYTHON_CMD. Right now this works because 3.7 is the default version of
Python 3 in the ports tree. However, at some point in the future, this
will be undoubtedly be changed.

It is also possible that users already might be building versions
newer than the default port/package, which needs to be an supported as
well.

In addition, the FDP recommends installing the 'python3' package. This
meta port/package depends on the default version of python, and
automatically creates a symlink to the default version.

With all this in mind, use '/usr/local/bin/python3' is used as the value
of PYTHON_CMD, since this means that the instructions will work
irrespective of which version is assigned.