Allow building the documentation and the website with upstream hugo
When we build our documentation and website we use a hacked version
of www/gohugo that allows for arbitrary asciidoc extensions.
This was added as a patch in r562416.

The suggested pull request[1] this patch was based on was not
accepted "as is" and only allows the execution of extensions in
Ruby's $LOAD_PATH.

With these changes, we add our <docroot>shared/lib directory to
Ruby's library directory so we can load our custom extensions. This
also allows building the docs and website outsite FreeBSD using
upstream and unmodified tools.

Tested with the upcoming vanilla gohugo-0.82 provided by
gohugo's maintainer.

[1] https://github.com/gohugoio/hugo/pull/8131

Patch by:               fernape@
Differential Revision:  https://reviews.freebsd.org/D29410