content/en: adjust paths to point at cgit.freebsd.org
An attempt has been made to provide the same kind of information as
before, but the layout and information displayed by cgit is different
from that of svnweb in some cases.

The reference to the commit log of Mk/bsd.port.subdir.mk did no longer
lead to a description of the supported parameters of "make search".
It has been replaced by an explicit description of all search modes.

Reviewed by:	wosch
Differential Revision:	https://reviews.freebsd.org/D29853