status: Announce the January-March 2021 status report
Because of a nit that was missed in the planning of the report, the URI
generation was not completely in-place when the status report was
initially commited, so the official announcement and email had to be
delayed a bit.

While here, make corrections to the README file, so that they reflect
reality.

Hat tip to:	lwhsu