documentation/Makefile: Add PDF targets.
Addition of various PDF targets allowing the build of books, articles
with the choice of the language.  For example, to build both en_US and
fr_FR books, use:
make DOC_LANG="en,fr" pdf-books
Documents are installed in documentation/public/lang/.  Targets to clean
the builds are also available.

Reviewed by:	ceri, gjb, dbaio
Differential Revision:	<https://reviews.freebsd.org/D30331>