documentation/Makefile: use all languages by default
Previously we only added languages with content in books/ to the
LANGUAGES variable because they would cause the build to fail, and
this meant that they did not get PDFs built either. It also meant they
could not be included in DOC_LANG as this then break the build.

Add a new variable, ARTICLEONLY_LANGS, to identify them and, with the
safety of the new logic in the tools/book-toc*py scripts to warn if we
miss one, build all of the content by default under the PDF target.

This also makes DOC_LANG safe to use even if there isn't a book in that
language.