Add Hugo's lang shortcode
This is a prerequisite for the translation workflow through .po files
and Weblate.

The asciidoctor include with the shortcode variable will help because
won't be needed to change the includes manually after converting the
translation (.po file) back to .adoc.

It will help other areas like, language attributes, which today is fixed
in every document.

The use of shortcode was suggested by jmooring@ from Hugo's project on
Github.

Reviewed by:	carlavilla
Differential Revision: https://reviews.freebsd.org/D30823