committer-guide/git: Add what to do for pushing main
git arc stage lands commits on main. Some people think it's more natural
to put changes on main and push that. Document how to do that as an
alternative to the somewhat safer way of rebasing a branch and pushing
that branch otno main. Also add tags to these sections so we (I?) can
reference them easily when people ask.

Sponsored by:		Netflix