Makefile: If DOC_LANG is set, stop hugo from building other languages.
The hugo driven targets currently build all the languages despite
anything set in DOC_LANG or elsewhere.

Have DOC_LANG drive HUGO_DISABLELANGUAGES in order to stop that.
With the caveat that: hugo always has to build en as it's the default
language, so add a warning if it's left out before adding it back in.

Reviewed by:	dbaio, ygy
Differential Revision:	https://reviews.freebsd.org/D30796