committers-guide: Document "Pull Request:" lines
A number of committers have started using "Pull Request:" to document
when a change was submitted by a pull request to give context discussed
there should it be necessary in the future. I documented it previously
in the `Landing a github pull request` section by implication as well.

Sponsored by:		Netflix