developers-handbook: update shared libraries policy section
Current text was written with a.out shared libraries system in mind,
which is not adequate for ELF system. In fact it can be considered
as mis-information causing to spread false rumors about FreeBSD
environment.

Drop everything that comes from a.out world, and briefly describe common
conventions related to ELF shared objects.

Reviewed by:	emaste
Sponsored by:	The FreeBSD Foundation
Differential revision:	https://reviews.freebsd.org/D31213