committers-guide: use "Unsupported" instead of "Tier 4"
Our use of "Tier 4" to represent unsupported architectures has caused
confusion, especially when we did not explain what the tiers mean on
in the platform list itself.  Just use "Unsupported" to describe the
architectures that are unsupported by a given release.

Reviewed by:	imp (earlier), ygy
Approved by:	ygy (mentor)
Sponsored by:	The FreeBSD Foundation
Differential Revision:	https://reviews.freebsd.org/D32030