Build offline documentation using Hugo and other improvements
* Allow to build the documentation offline using Hugo
* Fix link problems
* Remove SVN entities
* Fix lang problems to help weblate
* Move all the AsciiDoctor translations to the shared folder to
  help weblate
* Allow the possibility to build each document. For example each
document can be edited in and watched the result in real time in
VS Code or Vim/NeoVim.

Breaking changes:

Eliminate the possibility of building books or articles.
They must be builded at the same time.
This functionality will be restored when version 0.89.0 is in the
repositories in a couple of weeks.

Reviewed by:            dbaio@
Differential Revision:  https://reviews.freebsd.org/D31926