Convert ZFS chapter to active voice and remove weasel/unnecessary words
I used [1] to find the passive voice sentences and weasel words.
However, reviewers pointed out that some sentences were better of in the passive voice and that some of these weasel words were useful to understand the text. In those instances, I kept the original
sentences as they were.

The review became quite long in the process, but I think that the text
has improved a lot and that the concepts are better explained now.

Thanks to all the reviewers for their perseverance and good suggestions.

Reviewed by:		debdrup, ceri, ygy, pauamma_gundo.com
Differential Revision:	https://reviews.freebsd.org/D31707

[1] https://github.com/btford/write-good