Improve the website for browsers without Javascript enabled
Apply the light theme by default and then change it to the
theme selected by the user. This will allow users with a browser
with Javascript disabled use the Documentation Portal.

Also hide the theme chooser since it uses Javascript to change
the theme. In the future the property prefers-color-scheme
will be used to change the theme automatically to use the
browser theme.

PR:             261104
Submitted by:   Bob Frazier <bobf (AT) mrp3.com>