status/2022q1: Fix several small nits noticed late
A couple of reports had incorrect syntax for proper display, such as
having or not having the + indicator to indicate that the subsequent
text should be on the line below, as well as a few blank spaces that
made it past previous passes.
Also attempt to use manual page link generation where appropriate.

While here, convert everything to one sentence per line.