status/README: Add an order of operations
The order in which things get done isn't completely irrelevant, but it
wasn't very well-explained that there is an order let alone why it
should be followed.

I've added a series of steps of how I remember them needing to be
performed, but it's not meant to be exhaustive and can always be
modified at a later date.

I've also added a note about including a link to the rendered report in
the email; this suggestion was made to me by a few people, and makes a
lot of sense.

While here, I've also reflowed the file a bit, as well as replaced a few
words here and there to hopefully make it easier to understand, as well
as added an extra newline constant that accidentally got removed in one
of the report entries.