[hb][jails] Fix jail configuration example
jail.conf(8) accepts a devfs_ruleset that should be specified as a number
according to jail(8).

With the example in the handbook, we get this error when trying to launch the
jail:

Starting jails: cannot start jail  "www":
jail: www: devfs_ruleset: non-integer value "www_ruleset"

Remove that from the handbook and left the default value of 4, meaning
devfs_ruleset is enforced. No need to go into explaining devfs(8) at this point.

Reviewed by:	zlei.huang@gmail.com
Approved by:	0mp@
Differential Revision: https://reviews.freebsd.org/D35630