porters-handbook: fix example URLs
The format of URLs requested in this handbook had not been applied to
all the example URLs, e.g. with regard to preferring HTTPS over HTTP
and to add a skash "/" at the end of host and directory URLs.

Update the example URLs to follow the specified format rules, where
possible.