Update addkey.sh pointer in pgpkeys/README
Prior to the Hugo/AsciiDoctor conversion, the script lived in share/pgpkeys
next to the actual key files, but now it lives in a separate directory.
Though the committer's guide does list the full path, make this README more
standalone by providing breadcrumbs to indicate which script we're talking
about.