Fix URLs related to DigitalOcean
After DigitalOcean has discontinued FreeBSD as part of their supported
OS they have also removed some of the blog posts that refered to
FreeBSD. The linkw has been archived by the wayback machine. This patch
fixes those URL to point to the webarchive URL.

In addition to that one of the URL just above the links had a wrong
SiteURL pointing to digitalocean which was maybe due to someone copying
the previous definition. Fixed the SiteURL to point to
https://www.itwire.com

Part of the PR where digitalocean has to be removed from our supported
vendor has already been fixed.

PR:		264765
Approved by:	carlavilla (mentor)