handbooks: Remove references to portsnap
Although we are still generating portsnap deltas it is no longer the
recommended methods and git should be promoted. So remove references to
portsnap. Additionally

- Remove training whitespace
- Remove trailing consecutive blank lines

By no means this is an indication that portsnap is going to be removed
from the base or that portmgr@ is going to stop producing deltas. This
is just to demote the usages of portsnap and promote git.

PR:		264763
Approved by:	carlavilla (mentor)