books: Updating ports count stats in dev-model
Created a script that does the following:

- Calculates ports count yearly
- Calculates doc, ports, src, total unique committer numbers

This script is available at:
https://people.freebsd.org/~bofh/dropzone/dev-model.sh

And the graph generated is using this script. The tables has been
removed because it has been more than 28 years and putting this data in
the tables is not feasible at all. Additionally the tables of each 1000
milestone is difficult to achieve now hence has got ridden of it too.

Approved by:	bcr (mentor)
Differential Revision:	https://reviews.freebsd.org/D36972