license: Create a license guideline document
Create a license guideline document for the project. This pulls together
all the disparate docs into one location. This does not (yet) remove the
content from the other locations.

Also, introduce a new policy to allow SPDX identifiers to completely
specify a license and specify the mechanics of how to construe the
license for such files. This policy should be viewed as nearly complete
in spirit, but not yet finalized in wording. Ongoing legal review will
result in at least editorial changes.

Note 1: This changes the wording of our default license slightly in a way
that the SPDX legal team has deemed to be not legally different as
a practical matter in jurisdictions where Linux is relevant.

Note 2: This changes the default preferred license to detached, though
still allows the non-detached version.

Note 3: Text of the licenses is not (yet) in the src tree. That will
happend on the road to final approval.

I have one IP lawyer setup to review this, and invitations out to a few
more that specialize in the unique considerations that "open source IP"
and "open source projects" present (turns out they are hard to find and
often quite busy). Together, all these people should provide world-wide
coverage and may result in tweaks to account for variation in IP legal
regulatory regimes. Once this review is complete, I'll remove the
'other' copies of this information and pending final core approval it
will become the policy of the project. Getting it committed will elevate
its visibility and help me to get legal reviews more easily. Prior core
teams have approved this in spirit. The current core team will have
final approval once I've worked out the final rounds of feedback.

Approved in theory:	core.10 and core.11
Discussed with:		emaste, brooks, jhb, gnn
Sponsored by:		Netflix
Differential Revision:	https://reviews.freebsd.org/D29543