website/internal: Update outdated information
The build schedule is now simpler.

Remove the mirror page, which describes how to mirror the website, a
feature that is not working anymore, and there is no plan to make it
available again.

While here, also add a note about Varnish Cache that works with NGINX
and is widely used in the FreeBSD infrastructure.

Reviewed by:	lwhsu
Differential Revision: https://reviews.freebsd.org/D37492