Applications: first, the applications
Get straight to it! 

Applications first. Thousands of them, then use cases. 

Gut feeling: the section that's now at the foot of the page might be better elsewhere (a different page, or pages) but for now, let's not lose it. The first sentence in the section has a slightly 1990s taste to it.