git-primer: Update how to rebase a vender branch merge
Update the vendor branch merge section to include the missing commands
and hopefully to make it more clear. Switch to using bogus branch names
that will stand out on purpose.

This rework 'play' tested on the lua branch merge, which needed dozens
of rebasing...

Also, indent by one space the conflict section so as to not confuse
automated tools (or me) that it's a real conflict that needs to be
resolved.

Sponsored by:		Netflix