Trim a paragraph from the software license policy.
This bullet point seeks to make a broader point about the need for
compelling reasons for importing newer versions of 3rd party software
beyond just a version bump.

However, the license policy for source is not an ideal place to make
this point and it can be read as requiring core to validate updates to
any GPLv3 software in ports due to the surrounding context.

Reported by:	pauamma
Reviewed by:	grog, pauamma, imp, emaste
Differential Revision:	https://reviews.freebsd.org/D38131