Add OpenSSH updating note to 12.4 errata page
When updating from an earlier FreeBSD release (with an older OpenSSH
version) to 12.4 sshd may not accecpt new connections until it is
restarted.  Add a note to this effect, similar to 13.1.

This renders slightly strangely but I haven't been able to find out
why, so commit as is and we can tidy up later.

Sponsored by:	The FreeBSD Foundation
Differential Revision: https://reviews.freebsd.org/D38999