PackagesMacro: Extend the syntax to category/name@flavor[pkgname]
The package macro creates a link to the web interface of the port's
origin directory in the ports repository.  Extend its usage to cover the
case when we want to refer to a speicifed flavor of a port.

For example, if we want to refer the aarch64-gcc12 pacakge, in the past
we can only specify its origin port, like `pacakge:devel/freebsd-gcc12`,
and unable to add the `@aarch64` flavor suffix because it creates a
broken link.  This change makes it strip the suffix after `@` and accept
a parameter of package name as link text.  If not provided, use the origin
as the original behavior.

Thus we can have this kind of notation:

    package:devel/freebsd-gcc12@aarch64[aarch64-gcc12]

which creates linkt to:

    https://cgit.freebsd.org/ports/tree/devel/freebsd-gcc12/

with link text:

    aarch64-gcc12

Reviewed by:	carlavilla, dbaio
Differential Revision: https://reviews.freebsd.org/D39067