GSoC: Update and include wiki content
Much of the information on https://wiki.freebsd.org/SummerOfCode and
this page overlap.  Include all of the information here and retire the
wiki page.

While here, update the formatting so that it matches what has been
requested by doc committers in the past.

Approved by:	doc (carlavilla)
Sponsored by:	The FreeBSD Foundation
Differential Revision:	https://reviews.freebsd.org/D39187