porters-handbook: Remove shar(1) file submission
In contrast to the evolution of git there is no point in entertaining
shar files anymore. It was a good solution during the days of cvs/svn
but with git workflow this is counter productive for committers and
often the cause of INDEX break as the committers forget to add in the
relevant categories Makefile.

Approved by:	carlavilla(mentor) portmgr
Differential Revision:	https://reviews.freebsd.org/D39425