releng/: corrections, HTTPS for download-archive.⋯
MFC: fix the broken link.

I don't know why there's a bullet point for a single-item list, but for now there should be a space between the item and the subsequent indented paragraph.

Archived content is not limited to ISO files.

Optical media for content that is not archived by the Project is not limited to CD. The 10.0-RELEASE DVD at <https://www.freebsdmall.com/cgi-bin/fm/bsddvd10.0?id=XDwXP2ne&mv_pc=273>, and so on. 

More generally, the archive is not limited to disc images. 

Retailers (not vendors), for consistency with the FreeBSD Handbook. 

<https://reviews.freebsd.org/D39460#898329> prepare for configuration of  HTTPS for <https://download-archive.freebsd.org/>.