13.2R/readme: email, lists: corrections, updates
'help-there' was wrongly hyphenated. Fix this. 

<https://www.freebsd.org/mailman/listinfo> does not exist. Fix this, and the outdated mentions of Mailman. 

Remove the shouty box that tells everyone to not send subscription requests to the wrong addresses. This type of thing might be better placed in a book or article, although <https://lists.freebsd.org/> is clear enough for people to not make such mistakes. 

More broadly: update and improve this section of the page.