status: 2023q1: portmgr: corrections, suggestions
Fix both lists. 

The phrase 'ports on its dependencies' sounds wrong, I assume that it should be 'ports on their dependencies'. 

As noted in another recent PR, 'last quarter' is ambiguous; easily misinterpreted as 2022q4, which is not the subject of this (first) quarter's report. 

Character formatting for the four new USES, and bsd.ruby.mk.

Full stops for the final list item, where the first is preceded by a sentence with a colon.