status: 2023q1: portmgr: corrections, changes
Fix both lists.

The phrase 'ports on its dependencies' sounds wrong, I assume that it
should be 'ports on their dependencies'.

As noted in another recent PR, 'last quarter' is ambiguous; easily
misinterpreted as 2022q4, which is not the subject of this (first)
quarter's report.

Character formatting for the four new USES, for bsd.ruby.mk, and for
2022Q4.

A full stop for a final list item.

Approved by:  status (salvadore)
Pull request: https://github.com/freebsd/freebsd-doc/pull/172