vale/styles: Add BlockIgnore for [.programlisting]
In some older documents there are code snippets with the following tags
which are non standard for asciidoc:

[.programlisting]
....
CODE
....

As this is non-standard vale does not escape this and finds spelling
error but these blocks should be ignored by vale. This patch adds
relevant directives to ignore the blocks.

Approved by:	carlavilla (mentor)
Differential Revision:	https://reviews.freebsd.org/D39702