FreeBSD Handbook: Config… and Tuning: corrections
Minor corrections.

Names of some mailing lists lacked the 'the' prefix.

If referring to an outdated/legacy list archive, without mentioning
that it is not current: also, refer to the current archive.

One sentence per line should apply to most areas where a question mark
is followed by a space.