Web Resources: partial update
Begin to fix <https://bugs.freebsd.org/260968>, 

> Support: Web resources (page update)

FreeBSD News <https://www.truenas.com/> redirects to <https://www.truenas.com/>. Remove this item from the page. 

General UNIX® Information is not an ideal subheading for the various FreeBSD-specific links. I'm not rushing to create another subheading. 

A broader bug: heading level distinctions that are visible in markup, and in GitHub, are not visible at the FreeBSD site.