status: 2023q2: boot performance: improvements
If the links section is too narrow to state the title of the BSDCan
session, then state it elsewhere.

For consistency: efforts (plural).

I think, no need to repeat the full name beneath the sentence that
names Colin Percival (in full) as the sole contact.

Approved-by:  cperciva, salvadore
Pull-request: https://github.com/freebsd/freebsd-doc/pull/199