FreeBSD/ARM Project: clarity, order
The phrase 'ARM ports' is easily misinterpreted as ports for (not of) FreeBSD on ARM. 

Give prominence to the mailing list. 

Be concise. 

Make the status more positive; absences of releases and of packages are not a good beginning. Be true; FreeBSD/ARM can not be described as a (generally) excellent framework where there are neither releases, nor packages. Reorder and rewrite this section of the page – not only for positivity, also for clarity. 

For a newcomer who might want a brief introduction to tiering, immediate direction to the Committer's Guide is less than welcoming (thrown in at the deep end). Instead, ease the reader into the shallow end – the platforms page, which includes links to the Committer's Guide (for experienced swimmers). 

Make the hardware notes overview more readable. 

Begin tidying what needs to be done.