FreeBSD-related Publications: order, sanity, reality
Since the page gained a link to papers.freebsd.org, which is broader
than the bibliography in the FreeBSD Handbook, it no longer makes sense
to describe the bibliography as broad.

Reorder things – to no longer rely upon added emphasis (bold) to convey
what is, and is not, in the page.

https://www.startpage.com/do/dsearch?query=FreeBSD+publications
and comparable queries against other search engines have
FreeBSD-related Publications as the top hit, so:

- do not disappoint

- be upfront about the archaic essence of this publications page

- be upfront with directions to less archaic content.