Handbook: Internet resources: websites and more
Be concise, not repetitive. With 'Websites' as a subheading within the FreeBSD Handbook – and with Forums and FreeBSD within the name of The FreeBSD Forums – there's no need to describe The FreeBSD Forums as a web based discussion forum for FreeBSD … and so on.

Expand, reorder and rewrite the list of websites.

Internet Relay Chat is a communication protocol that does not use HTTP or HTTPS; it's not a website. Make it separate.

Below the separate section for Internet Relay Chat, add a hidden section in readiness for Matrix.