status: 2023q2: KDE: markup
Two sentences per line may be excessive in contexts such as AsciiDoc.

Whilst it's sometimes true that too much is never enough,
multiple colons may be excessive in a package:⋯/⋯[] context for a
port that is not necessarily packaged. In this case,
package:ports-mgmt/packagekit-qt[] port, the
package ports management package kit cutie port is packaged.

Fixes: b2678db626fb Status/2023Q2/kde.adoc: Add report
Fixes: 10374142a520 status: 2023q2: KDE: fixes and minor changes