.vale/styles: Removing Foreign Language references from English contents
e.g/eg or ie/i.e actually hails from Latin root and do not have any
specific general acceptance in any English standards or at least should
not be written in technical documentation. But we have a lot of these in
our documentation. Throw errors for such occurrences and replace those
with for example or that is.

Approved by:	carlavilla