Commercial Vendors: reorder, reformat, improve
Don't begin by linking to a page that lacks descriptions. 

Be calm, be correct. A person seeking a commercial vendor is not likely to be excited by 'Books and accessories you just cannot live without!' where only three things are listed, and the first is neither a book nor an accessory! Life goes on! Use fewer exclamation marks! This is not (UK tabloid) The Sun!

Grammar, for example: 

- punctuation missing from things such as 'Miscellaneous Vendors Books and accessories you …'
- duplication in 'Software Vendors Vendors From audio players, …'
- 'supports a … support'
- superfluous 'our' in 'our endorsement of their products or services by the FreeBSD Project.'

Some attention to weasel words …

… and so on (expect further improvements in and around this page).