consult_bycat: deduplicate
Weasel word: growing. (Not always; lists sometimes shrink.)

What was (or is) stated at the parent page – note, the sidebar and menu contexts for https://www.freebsd.org/commercial/ – need not be restated, verbatim or condensed, in pages such as these. 

Aiming to have the same text at eight or more subpages is not only a maintenance burden, there's also resultant nonsense.