Handbook - Security: document OpenSSL's FIPS provider
The new FIPS provider module from OpenSSL 3 is protected from use
without passing an integrity check, which is expected to be configured
by the local system administrator.

This new section documents a way to configure this on FreeBSD, so that
every user on the system is allowed to pass this check and use the FIPS
module.