Status: 2023q3: Core: corrections, changes
Grammar: a person (singular) can be alumna or alumnus, but not alumni
(plural).

Project (not Project in Coimbra).

Uppercase P for Project (the FreeBSD Project).

Reduce repetition.

If Matrix is to be described in a single word, do not use the word
'IM'.  It is not simply text-based instant messaging.  

https://matrix.org/about/