Status: 2023q3: doceng: corrections, markup
This quarter, not last.

Add missing markup, for list items.  Compare with
https://www.freebsd.org/status/report-2023-01-2023-03/#_documentation_engineering_team

Real names for people who requested safekeeping of their commit bits.
Compare with the same past report.

Include the resignation of my doc commit bit.  Express this in the same
way as the safekeeping of Sergey Kandaurov's doc commit bit.