FAQ: Complete FAQ rewrite
Complete rewrite of the FAQ is being carried out to offer a version
agnostic to the FreeBSD version and some simple entries.

The idea behind this rewrite has been to simplify the FAQ and make it
more accessible.

The FAQ is an entry point to FreeBSD, for more complete questions we
have the Handbook.

PR:                     266111, 263571, 273313, 273313, 261765, 248351, 261816
Reviewed by:            acm@, ashish@, cc@, emaste, fernape@, linimon@
Differential Revision:  https://reviews.freebsd.org/D41644
Sponsored by:           Daifressh