SA/EN templates: drop freebsd-update i386 support for 14.x+
In 2021 we announced[1] that i386 was demoted to Tier 2 status for
FreeBSD 13.x.  We stated that i386 would receive support beyond what is
specified by Tier 2 status and would receive binary updates for 13.x, as
an exception to the standard Tier policy.

We also stated that "i386 will receive reduced support inline with
existing Tier 2 platforms for 14.x".  So, now document that binary
updates for i386 apply to 13.x and lower.

There is a small nit here, in that we don't officially support arm64 on
FreeBSD 12 in freebsd-update but have removed that caveat in this change.

My original proposed text was

    Systems running a RELEASE version of FreeBSD on the amd64, (on
    FreeBSD 13 and earlier) i386, or (on FreeBSD 13 and later) arm64
    platforms can be updated ...

gordon@ pointed out that this text is rather awkward.

We did actually build 12.4 arm64 updates in the last set of SA/ENs, and
the branch is EOL at the end of the year anyway.  So, overlook the
12.x/arm64 quirk and document just the i386 exception.

[1] https://lists.freebsd.org/pipermail/freebsd-announce/2021-January/002006.html

Reviewed by:	delphij
Sponsored by:	The FreeBSD Foundation
Differential Revision: https://reviews.freebsd.org/D42506