SA/EN templates: clarify referenced hashes / revisions
A fix may arrive in stable branches as multiple commits (e.g., if there
is an issue with the initial change).  Update the templates to mention
that the issue is corrected "as of" the specified hash rather than "by"
the specified hash.

Reported by:	brooks
Reviewed by:	delphij, brooks
Sponsored by:	The FreeBSD Foundation
Differential Revision: https://reviews.freebsd.org/D42507