bsdinstall: modernize the screenshots
This contains a new version of every screenshot still relevant to the
bsdinstall section of the FreeBSD handbook. They were made after booting
in EFI mode instead of the traditional BIOS mode, for a more modern look
as well as a more accurate resemblance to the actual look & feel of the
installer on current hardware.

Sponsored by:	The FreeBSD Foundation