Update _index.adoc
I found a this section to be unclear, so I have explained it more clearly. I have also changed the reference 'xorg' to 'the X Window System' as this is what the chapter is about. Calling an app by its package name is lazy and could be confusing to someone who hasn't had any experience with the X Window System.