Fix thinjail instructions on UFS
The zfs instructions say to clone a snapshot from the skeleton to the new thinjail. But the UFS instructions say to _make_ the thinjail directory first and then copy the skeleton into it. This will cause your new jail to be inside of `/usr/local/jails/containers/thinjail/13.2-RELEASE-skeleton` instead of just `/usr/local/jails/containers/thinjail`.